+++ app/helpers/match_helper.rb
@@ -1,0 +1,25 @@
+module MatchHelper
+  def glyphicon
+    case status
+      when 'scheduled'
+        "glyphicon-pencil"
+      when 'started'
+        "glyphicon-ok"
+      when 'ended'
+        "glyphicon-flag"
+      when 'cancelled'
+        "glyphicon-remove"
+    end
+  end
+
+  def getTimeDifference(time)
+    delta = (time - Time.now).abs
+
+    minutes = (delta / 60) % 60
+    hours = (delta / (60 * 60)) % 24
+    days = (delta / (60 * 60 * 24))
+
+    time > Time.now ? format("%dd %dh %dm", days, hours, minutes) : format("-(%dd %dh %dm)", days, hours, minutes)
+  end
+
+end
