+++ app/models/match.rb
@@ -16,28 +16,6 @@ class Match < ActiveRecord::Base
   scope :overdue, -> { Match.scheduled.where('starts_at < ?', Time.now) }
   enum status: { scheduled: 0, started: 1, ended: 2, cancelled: 3 }
 
-  def glyphicon
-    case status
-      when 'scheduled'
-        "glyphicon-pencil"
-      when 'started'
-        "glyphicon-ok"
-      when 'ended'
-        "glyphicon-flag"
-      when 'cancelled'
-        "glyphicon-remove"
-    end
-  end
-
-  def getTimeDifference(time)
-    delta = (time - Time.now).abs
-
-    minutes = (delta / 60) % 60
-    hours = (delta / (60 * 60)) % 24
-    days = (delta / (60 * 60 * 24))
-
-    time > Time.now ? format("%dd %dh %dm", days, hours, minutes) : format("-(%dd %dh %dm)", days, hours, minutes)
-  end
 
   attr_accessor :png, :ball, :players_a, :playersB, :img_counter, :goalie_a, :goalie_b
   attr_accessor :all_players
